# swagger.json
{
    "openapi": "3.0.1",
    "info":{
        "title": "CRUD Users",
        "description": "Documentação do CRUD de usuários",
        "version": "1.0.0"
    },
    "basePath": "/",
    "paths":{
        "/userscrud/register":{
            "post": {
                "tags":["Route Users"],
                "requestBody":{
                        "content":{
                            "application/json":{
                                "schema":{
                                    "type": "object",
                                    "properties":{
                                        "name":{
                                            "type":"string"
                                        },
                                        "surname":{
                                            "type":"string"
                                        },
                                        "password":{
                                            "type":"string"
                                        },
                                        "confirmPassword":{
                                            "type":"string"
                                        },
                                        "phone":{
                                            "type":"string"
                                        },
                                        "cpf":{
                                            "type":"string"
                                        }
                                    },
                                    "example":{                                   
                                        "name": "Pedro",
                                        "surname": "Silva",
                                        "phone": "(11)99938-3221",
                                        "cpf": "885.226.339-10",
                                        "password": "123456",
                                        "confirmPassword":"123456"
                                    }

                                }

                            }
                        }
                },         
                "responses":{
                    "400":{
                        "description": "Formato dos campos informados inválido"
                    },
                    "401":{
                        "description": "Usuário não autorizado a acessar essa rota"
                    },
                    "200":{
                        "description": "Usuário criado no banco"
                    }
                }
            }
        },
        "/userscrud/authenticate":{
            "post": {
                "tags":["Route Users"],
                "requestBody":{
                        "content":{
                            "application/json":{
                                "schema":{
                                    "type": "object",
                                    "properties":{
                                        "cpf":{
                                            "type":"string",
                                            "required": true
                                        },                      
                                        "password":{
                                            "type":"string"
                                        }                            
                                    },
                                    "example":{                                   
                                        "cpf": "885.226.339-10",
                                        "password": "123456"
                                    }

                                }

                            }
                        }
                },         
                "responses":{
                    "400":{
                        "description": "A senha informada está incorreta/CPF não encontrado na base de dados"
                    },
                    "200":{
                        "description": "Login efetuado com sucesso."
                    }
                }
            }
        },
        "/userscrud/update":{
            "put": {
                "tags":["Route Users"],
                "requestBody":{
                    "content":{
                        "application/json":{
                            "schema":{
                                "type": "object",
                                "properties":{
                                    "userId":{
                                        "type":"string",
                                        "required": true
                                    },
                                    "name":{
                                        "type":"string"
                                    },
                                    "surname":{
                                        "type":"string"
                                    },
                                    "phone":{
                                        "type":"string"
                                    },
                                    "cpf":{
                                        "type":"string",
                                        "required":true
                                    },
                                    "password":{
                                        "type":"string",
                                        "required":true
                                    },
                                    "confirmPassword":{
                                        "type":"string",
                                        "required":true
                                    }
                                },
                                "example":{                                                                 
                                        "userId":"62605b3fbcbe86bc08cc3465",
                                        "name": "Jones",
                                        "surname": "Maikon",
                                        "password":"1234567",
                                        "confirmPassword":"1234567",
                                        "phone": "(49)99912-6595",
                                        "cpf": "047.002.619-97"                                  
                                }
                            }

                        }
                    }
            },
                "responses":{
                    "400":{
                        "description": "Ocorreu um erro na atualização dos dados."
                    },
                    "401":{
                        "description": "Usuário não autorizado a acessar essa rota"
                    },
                    "200":{
                        "description": "Dados atualizados com sucesso."
                    }   
                }
            }
        },
        "/userscrud/list":{
            "get": {
                "tags":["Route Users"],
                "parameters": [
                    {
                        "in": "query",
                        "name": "cpf",
                        "type": "string"
                    }
                ],
                "responses":{
                    "400":{
                        "description": "Informações de CPF não armazenadas."
                    },
                    "401":{
                        "description": "Usuário não autorizado a acessar essa rota"
                    },
                    "200":{
                        "description": "Dados encontrados."
                    }   
                }
            }
        },
        "/userscrud/remove":{
            "delete": {
                "tags":["Route Users"],
                "parameters": [
                    {
                        "in": "query",
                        "name": "userId",
                        "type": "string",
                        "required":true
                    }
                ],
                "responses":{
                    "400":{
                        "description": "Ocorreu um erro na exclusão do usuário"
                    },
                    "401":{
                        "description": "Usuário não autorizado a acessar essa rota"
                    },
                    "200":{
                        "description": "Usuário excluído com sucesso."
                    }   
                }
            }
        }
    }

}